moving settings files to specific package
